fixing data types from database

+++ database/migrations/20190415222817_kitchensTable.js
@@ -13,7 +13,7 @@ exports.up = function(knex, Promise) {
       .string("mission", 300)
       .notNullable()
       .unique();
-    table.integer("average_visitors");
+    table.float("average_visitors");
     table.string("website");
   });
 };
